+++ lib/flutter_arc_text.dart
@@ -1,3 +1,5 @@
 library flutter_arc_text;
 
 export 'src/arc_text.dart';
+export 'src/arc_text_painter.dart';
+export 'src/enums.dart';