+++ lib/src/enums.dart
@@ -1,0 +1,3 @@
+enum StartAngleAlignment { start, center, end }
+enum Direction { clockwise, counterClockwise }
+enum Placement { inside, outside }